html/browsers/browsing-the-web/navigating-across-documents/015.html

<!doctype html>
<!--/
specs:
  - id: browsing-the-web.html#navigating-across-documents
    steps: [12]
  - id: browsing-the-web.html#javascript-protocol
  - id: text-level-semantics.html#the-a-element
  - id: webappapis.html#the-event-handler-processing-algorithm
  - id: iframe-embed-object.html#the-iframe-element:about:blank

description: |
  1. This test creates an anchor with an href containing a javascript: url.
  2. The test "clicks" the anchor, triggering the javascript: href navigation,
      which must be evaluated and executed asynchronously.
  3. location.href is used to prove that the `javascript:` url's contents
      are executed in the iframe.

/-->
<title>Link with onclick and href javascript: url, order of execution</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<a target="test" onclick="parent.reports.push('click');" href="javascript:parent.reports.push(location.href);">Test</a>
<script>
var t = async_test(undefined, {
  timeout:4000
});
var reports = [];
t.step(function() {
  document.getElementsByTagName("a")[0].click();
  reports.push("after click");
});
onload = t.step_func(function() {
  // javascript: executions are async.
  assert_array_equals(reports, ["click", "after click", "about:blank"]);
  t.done();
});
</script>
