LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/015-expected.txt


FAIL  Link with javascript onclick and href script order  assert_array_equals: property 1, expected "after script" but got "href"
 Test
